% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\name{compare_documentsnearest}
\alias{compare_documentsnearest}
\title{Compares a document to all other documents in the history in order to find a matching cluster}
\usage{
compare_documentsnearest(document_vector, all_documents)
}
\arguments{
\item{document_vector}{TDT-weighted representation of the document about to add.}

\item{all_documents}{TDT-weighted representation of the history.}
}
\value{
A indicator of a matched document in the history or NULL if threshold was not exceeded.
}
\description{
Compares a document to all other documents in the history in order to find a matching cluster
}
\examples{
compare_documentsnearest(new_weighted_dtm[i,,drop=F], o1@history)
}
