% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\name{create_documents}
\alias{create_documents}
\title{Creates new TDT compatibe documents. This introduces a term weighting to the new document-term-matrix}
\usage{
create_documents(new_dtm, document_frequency, N)
}
\arguments{
\item{new_dtm}{The added document-term-matrix}

\item{document_frequency}{The document frequency histogram to calculate the weigthing}

\item{N}{The number of all documents included in the process (history + new_dtm)}
}
\value{
A document-term-matrix containing the largest 1000 term weights calculated from new_dtm.
}
\description{
Creates new TDT compatibe documents. This introduces a term weighting to the new document-term-matrix
}
\examples{
new_weighted_dtm <- create_documents(new_dtm,  o1@document_frequency, nrow(o1@history) + nrow(new_dtm))
}
