% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\name{get_weight}
\alias{get_weight}
\title{Calculates the term-weighting according to TDT. It is basically a TF-IDF weighting scheme.}
\usage{
get_weight(new_dtm, df, N)
}
\arguments{
\item{new_dtm}{The document-term-matrix to be processed.}

\item{df}{The document frequency information.}

\item{N}{The number of the documents.}
}
\value{
The weighted document-term-matrix
}
\description{
Calculates the term-weighting according to TDT. It is basically a TF-IDF weighting scheme.
}
