% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Adds the += operator to add a vector to another}
\usage{
e1 \%+=\% e2
}
\arguments{
\item{e1}{First vector}

\item{e2}{Second vector}
}
\value{
The added value vector
}
\description{
Adds the += operator to add a vector to another
}
\examples{
v1 <- c(1,2,3,4)
v2 <- c(2,3,4,5)
v1 \%+=\% v2
v1
}
