% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\docType{methods}
\name{initialize,tdt_tracker-method}
\alias{initialize,tdt_tracker-method}
\title{Initializes a tdt_tracker S4 object.}
\usage{
\S4method{initialize}{tdt_tracker}(.Object, ...)
}
\value{
The initialized tdt_tracker object.
}
\description{
Initializes a tdt_tracker S4 object.
}
