% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\name{sim_ab}
\alias{sim_ab}
\title{Cosine similarity for 2 documents}
\usage{
sim_ab(a, b)
}
\arguments{
\item{a}{Vector of document A.}

\item{b}{Vector of document B.}
}
\value{
Distance between 2 documents.
}
\description{
Cosine similarity for 2 documents
}
