% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\docType{class}
\name{tmodel}
\alias{tmodel}
\title{A Topic Model Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
#See available algorithms
tmodel$list_methods()
#Create a new Instance. Specify the algorithm by putting in one of the strings from the output of the previous command.
t <- tmodel$new(method = "topicmodels::LDA")
#Input a textcorpus. This should contain a column with the name "text". The rows represent the documents.
t$input( datasource = reuters_training_data)
# Now you can create a topic model based on the input data
t$createTM()
#After that you can infer topic distribution for unseen documents with:
$inferTopic(reuters_test_data)
#Or you can try to evaluate the current topicmodel by starting:
t$evaluate()
}
\description{
An R6 class to execute, evaluate and refine a topic model independently of the specific underlying algorithm.
             This uses an implementation of an algorithm as specified in the documentation of tm_abstr as engine.
             You can execute the methods to infer a topic model, start an evaluation task and refine parameters, or classify unseen Text.
             The algorithm to be used can be set at initialization time. See usage examples.
}
\section{Methods}{

\describe{
 \item{\code{set_parameters(par_list)}}{
       This method uses a List of named Parameters to set the appropriate values.
       The Parameters needed/available will be checked by the specific implementation.
 }
 \item{\code{get_parameters()}}{
        Get a list of available parameters.
 }
 \item{\code{get_model()}}{
       return the model parameters
       List containing named elements "phi" (topic-word matrix), and "theta" (document-topic matrix).
 }
 \item{\code{input(datasource = stop("Need text to input"))}}{
        Data input method. \code{datasource} should contain a column named "text". Each row will be treated as one document.
 }
 \item{\code{createTM()}}{
        Start estimating model parameters.
 }
 \item{\code{evaluate()}}{
       Start an interactive evaluation task on the currently estimated model.
       Specifically this will measure topic coherence, topic intrusion, word instrusion, and reproducability.
 }
 \item{\code{inferTopics(new_text = stop("Need text to infer topics"))}}{
       Infer a topic distribution with the current model for unseen text. new_text should be in the same format as \code{datasource} in \code {input()}
 }
}
}

\keyword{data}
