% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\docType{data}
\name{topicmodels_lda}
\alias{topicmodels_lda}
\title{Implementation of topicmodels::LDA algorithm}
\format{R6Class}
\usage{
topicmodels_lda
}
\description{
Implementation of the topicmodel::LDA algorithm. Implements all of the baseclass methods.
             Parameters are the number of topics K and the prior alpha
}
\section{Slots}{

\describe{
\item{\code{name}}{Sets the name of the method/function (identifier for the class)}

\item{\code{package}}{Sets the name of the package the method was imported from}

\item{\code{parameters}}{Named list of available parameters: alpha and K. This list determines which values CAN be set and also saves their value.}
}}

\section{Methods}{

\describe{
  \item{\code{call()}}{This should execute the algorithm and save its results to the internal variable (private$model)}
  \item{\code{input(dtm = NULL)}}{ Input the training DTM, transform to the algorithm specific format and save to private$internal_representation.}
  \item{\code{output()}}{Tranform the internal representation}
  \item{\code{inferTopics(new_text=NULL)}}{Infer the topicdistribution with regards to the calculated model for yet unseen data.
              Return the document x topic matrix containing the probabilities of assignment. }
  \item{\code{set_parameters(par_list)}}{This method uses a List of named Parameters to set the appropriate values of the implemented algorithm.
                                         It is implemented in this class. It checks whether the input parameters are available in the implementation.
                                         It only sets the parameters if there are no mismatches. }
   \item{\code{info()}}{Returns the package and algorithm name.}
}
}

\author{
Janos Borst
}
\keyword{datasets}
