% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cluster_tdt.R
\name{update_document_frequency}
\alias{update_document_frequency}
\title{Takes a vector of document frequencies and adds up new document frequency counts. Also aligns vocabulary to new words.}
\usage{
update_document_frequency(document_frequency_vector,
  document_frequency_vector_update)
}
\arguments{
\item{document_frequency_vector}{Old document frequency vector}

\item{document_frequency_vector_update}{New document frequency vector}
}
\value{
The updated vector
}
\description{
Takes a vector of document frequencies and adds up new document frequency counts. Also aligns vocabulary to new words.
}
\examples{
o1@document_frequency <- update_document_frequency(
 o1@document_frequency,
 colSums(tmca.util::make_binary(new_dtm))
)
}
